requirements.txt
mysql-connector-python == 9.5.0
python-json-logger == 3.3.0
boto3 == 1.40.45